Made Player scene Autoload and fixed tests
Set the Player scene as an autoload. This way it is constantly loaded and can always be accesed by all nodes.
ProcessMode and UI Visibility are only true if the current menu of the MenuManager is set to main_game.

Also fixed tests so that they use the new structure of the game